refactor: move config tests to separate file
- Moved all config tests from src/config.rs to tests/config_test.rs
- Reduced src/config.rs line count below file length limit
- Maintained all existing tests and functionality
- Tests for OpenAI config loading, defaults, and validation included in the new location
