+++ benches/bench_utils.rs
@@ -44,6 +44,9 @@ pub fn setup_logging(
                 port: "0".to_string(),                     // Not used in benchmarks
                 anthropic_api_key: "test-key".to_string(), // Not used in benchmarks
                 anthropic_target_url: "https://example.com".to_string(), // Not used in benchmarks
+                openai_api_key: None,                      // Not used in benchmarks
+                openai_api_base_url: "https://api.openai.com".to_string(), // Not used in benchmarks
+                openai_enabled: false,                     // Not used in benchmarks
                 log_stdout_level: "debug".to_string(),
                 log_format: "json".to_string(), // JSON format for consistency
                 log_bodies: true,
@@ -67,6 +70,9 @@ pub fn setup_logging(
                 port: "0".to_string(),                     // Not used in benchmarks
                 anthropic_api_key: "test-key".to_string(), // Not used in benchmarks
                 anthropic_target_url: "https://example.com".to_string(), // Not used in benchmarks
+                openai_api_key: None,                      // Not used in benchmarks
+                openai_api_base_url: "https://api.openai.com".to_string(), // Not used in benchmarks
+                openai_enabled: false,                     // Not used in benchmarks
                 log_stdout_level: "off".to_string(),       // Disable stdout logging
                 log_format: "json".to_string(),            // Not relevant when stdout disabled
                 log_bodies: true,
@@ -90,6 +96,9 @@ pub fn setup_logging(
                 port: "0".to_string(),                     // Not used in benchmarks
                 anthropic_api_key: "test-key".to_string(), // Not used in benchmarks
                 anthropic_target_url: "https://example.com".to_string(), // Not used in benchmarks
+                openai_api_key: None,                      // Not used in benchmarks
+                openai_api_base_url: "https://api.openai.com".to_string(), // Not used in benchmarks
+                openai_enabled: false,                     // Not used in benchmarks
                 log_stdout_level: "debug".to_string(),
                 log_format: "json".to_string(), // JSON format for consistency
                 log_bodies: true,
