+++ src/config.rs
@@ -170,6 +170,7 @@ pub struct Config {
 
     // OpenAI API configuration
     /// API key for authenticating with OpenAI API (None if disabled)
+    #[allow(dead_code)] // Will be used when OpenAI proxy handler is implemented
     pub openai_api_key: Option<String>,
     /// Target URL for the OpenAI API
     pub openai_api_base_url: String,
@@ -417,347 +418,5 @@ pub fn load_config() -> &'static Config {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use std::collections::HashMap;
-    use std::env;
-    use std::sync::Mutex;
-    use std::sync::Once;
-
-    // Use a mutex to ensure environment variable tests don't interfere with each other
-    static ENV_MUTEX: Mutex<()> = Mutex::new(());
-    static INIT: Once = Once::new();
-
-    // Initialize test environment exactly once
-    fn initialize() {
-        INIT.call_once(|| {
-            // Initialize test environment here
-        });
-    }
-
-    // A function to create a test config with specific environment variables
-    fn create_test_config_with_env(env_vars: HashMap<&str, &str>) -> Config {
-        // Ensure synchronization across tests
-        let _lock = ENV_MUTEX.lock().unwrap();
-        initialize();
-
-        // Save current environment
-        let mut old_vars = HashMap::new();
-        for (key, _) in env_vars.iter() {
-            old_vars.insert(*key, env::var(*key).ok());
-        }
-
-        // Set provided environment variables
-        for (key, value) in env_vars.iter() {
-            // Only set non-empty environment variables
-            if !value.is_empty() {
-                env::set_var(key, value);
-            } else {
-                env::remove_var(key);
-            }
-        }
-
-        // Create the config (similar to create_test_config but cleaner)
-        let port = env::var("PORT").unwrap_or_else(|_| DEFAULT_PORT.to_string());
-        let anthropic_api_key =
-            env::var("ANTHROPIC_API_KEY").unwrap_or_else(|_| "test-api-key".to_string());
-        let anthropic_target_url = env::var("ANTHROPIC_TARGET_URL")
-            .unwrap_or_else(|_| DEFAULT_ANTHROPIC_TARGET_URL.to_string());
-        let log_stdout_level =
-            env::var("LOG_LEVEL").unwrap_or_else(|_| DEFAULT_LOG_STDOUT_LEVEL.to_string());
-        let log_format = env::var("LOG_FORMAT").unwrap_or_else(|_| DEFAULT_LOG_FORMAT.to_string());
-        let log_bodies = env::var("LOG_BODIES")
-            .map(|v| v.to_lowercase() != "false" && v != "0")
-            .unwrap_or(DEFAULT_LOG_BODIES);
-        let log_file_path =
-            env::var("LOG_FILE_PATH").unwrap_or_else(|_| DEFAULT_LOG_FILE_PATH.to_string());
-        let log_file_level =
-            env::var("LOG_FILE_LEVEL").unwrap_or_else(|_| DEFAULT_LOG_FILE_LEVEL.to_string());
-        let log_max_body_size = env::var("LOG_MAX_BODY_SIZE")
-            .ok()
-            .and_then(|size_str| {
-                size_str.parse::<usize>().ok().or_else(|| {
-                    eprintln!(
-                        "Failed to parse LOG_MAX_BODY_SIZE: '{}', using default {}",
-                        size_str, DEFAULT_LOG_MAX_BODY_SIZE
-                    );
-                    None
-                })
-            })
-            .unwrap_or(DEFAULT_LOG_MAX_BODY_SIZE);
-
-        // Parse LOG_DIRECTORY_MODE
-        let log_directory_mode = env::var("LOG_DIRECTORY_MODE")
-            .map(|mode| match mode.to_lowercase().as_str() {
-                "xdg" => LogDirectoryMode::Xdg,
-                "system" => LogDirectoryMode::System,
-                _ => LogDirectoryMode::Default,
-            })
-            .unwrap_or(LogDirectoryMode::Default);
-
-        let config = Config {
-            port,
-            anthropic_api_key,
-            anthropic_target_url,
-            openai_api_key: None,
-            openai_api_base_url: DEFAULT_OPENAI_TARGET_URL.to_string(),
-            openai_enabled: DEFAULT_OPENAI_ENABLED,
-            log_stdout_level,
-            log_format,
-            log_bodies,
-            log_file_path,
-            log_file_level,
-            log_max_body_size,
-            log_directory_mode,
-            log_max_age_days: None,
-        };
-
-        // Restore old environment
-        for (key, value_opt) in old_vars {
-            match value_opt {
-                Some(value) => env::set_var(key, value),
-                None => env::remove_var(key),
-            }
-        }
-
-        config
-    }
-
-    #[test]
-    fn test_default_values() {
-        // For default values, we just need the API key (required) and all others empty/unset
-        let mut env_vars = HashMap::new();
-        env_vars.insert("ANTHROPIC_API_KEY", "test-api-key");
-
-        // These variables should be unset for default tests, not empty strings
-        let vars_to_unset = vec![
-            "PORT",
-            "ANTHROPIC_TARGET_URL",
-            "LOG_LEVEL",
-            "LOG_FORMAT",
-            "LOG_BODIES",
-            "LOG_FILE_PATH",
-            "LOG_FILE_LEVEL",
-            "LOG_MAX_BODY_SIZE",
-        ];
-
-        // Explicitly remove them from the environment
-        for var in vars_to_unset {
-            env::remove_var(var);
-        }
-
-        let config = create_test_config_with_env(env_vars);
-
-        // Verify default values
-        assert_eq!(config.port, "8080");
-        assert_eq!(config.anthropic_api_key, "test-api-key");
-        assert_eq!(config.anthropic_target_url, "https://api.anthropic.com");
-        assert_eq!(config.log_stdout_level, "info");
-        assert_eq!(config.log_format, "pretty");
-        assert!(config.log_bodies);
-        assert_eq!(config.log_file_path, "./switchboard.log");
-        assert_eq!(config.log_file_level, "debug");
-        assert_eq!(config.log_max_body_size, 20480);
-        assert_eq!(config.log_directory_mode, LogDirectoryMode::Default);
-    }
-
-    #[test]
-    fn test_environment_variable_parsing() {
-        let env_vars = HashMap::from([
-            ("PORT", "9090"),
-            ("ANTHROPIC_API_KEY", "custom-api-key"),
-            ("ANTHROPIC_TARGET_URL", "https://custom.example.com"),
-            ("LOG_LEVEL", "debug"),
-            ("LOG_FORMAT", "json"),
-            ("LOG_BODIES", "false"),
-            ("LOG_FILE_PATH", "/tmp/custom.log"),
-            ("LOG_FILE_LEVEL", "trace"),
-            ("LOG_MAX_BODY_SIZE", "10240"),
-            ("LOG_DIRECTORY_MODE", "xdg"),
-        ]);
-
-        let config = create_test_config_with_env(env_vars);
-
-        // Verify custom values were used
-        assert_eq!(config.port, "9090");
-        assert_eq!(config.anthropic_api_key, "custom-api-key");
-        assert_eq!(config.anthropic_target_url, "https://custom.example.com");
-        assert_eq!(config.log_stdout_level, "debug");
-        assert_eq!(config.log_format, "json");
-        assert!(!config.log_bodies);
-        assert_eq!(config.log_file_path, "/tmp/custom.log");
-        assert_eq!(config.log_file_level, "trace");
-        assert_eq!(config.log_max_body_size, 10240);
-        assert_eq!(config.log_directory_mode, LogDirectoryMode::Xdg);
-    }
-
-    #[test]
-    fn test_boolean_parsing() {
-        // Test various boolean string representations
-        let test_cases = vec![
-            ("true", true),
-            ("True", true),
-            ("TRUE", true),
-            ("1", true),
-            ("yes", true),
-            ("Y", true),
-            ("false", false),
-            ("False", false),
-            ("FALSE", false),
-            ("0", false),
-            ("no", true), // This should be true since we only check for "false" and "0"
-            ("n", true),  // Same here
-        ];
-
-        for (input, expected) in test_cases {
-            let mut env_vars = HashMap::new();
-            env_vars.insert("ANTHROPIC_API_KEY", "test-api-key");
-            env_vars.insert("LOG_BODIES", input);
-
-            let config = create_test_config_with_env(env_vars);
-            assert_eq!(config.log_bodies, expected, "Failed for input: {}", input);
-        }
-    }
-
-    #[test]
-    fn test_numeric_parsing_valid() {
-        let env_vars = HashMap::from([
-            ("ANTHROPIC_API_KEY", "test-api-key"),
-            ("LOG_MAX_BODY_SIZE", "12345"),
-        ]);
-
-        let config = create_test_config_with_env(env_vars);
-        assert_eq!(config.log_max_body_size, 12345);
-    }
-
-    #[test]
-    fn test_numeric_parsing_invalid() {
-        let env_vars = HashMap::from([
-            ("ANTHROPIC_API_KEY", "test-api-key"),
-            ("LOG_MAX_BODY_SIZE", "not-a-number"),
-        ]);
-
-        let config = create_test_config_with_env(env_vars);
-        assert_eq!(config.log_max_body_size, 20480);
-    }
-
-    #[test]
-    fn test_edge_case_large_value() {
-        let max_size_str = usize::MAX.to_string();
-        let env_vars = HashMap::from([
-            ("ANTHROPIC_API_KEY", "test-api-key"),
-            ("LOG_MAX_BODY_SIZE", max_size_str.as_str()),
-        ]);
-
-        let config = create_test_config_with_env(env_vars);
-
-        // The behavior is platform-dependent, so we should check that:
-        // 1. Either the value is correctly parsed as usize::MAX, OR
-        // 2. The value is at least the default (20480) and not something invalid
-        assert!(config.log_max_body_size == usize::MAX || config.log_max_body_size >= 20480);
-    }
-
-    #[test]
-    fn test_empty_string_environment_variable() {
-        // In Rust, setting an environment variable to an empty string with env::set_var
-        // is equivalent to removing it for env::var (returns Err)
-        // Our test utility now removes empty string vars to match this behavior
-        let mut env_vars = HashMap::new();
-        env_vars.insert("ANTHROPIC_API_KEY", "test-api-key");
-
-        // First test with the variable unset
-        let config = create_test_config_with_env(env_vars.clone());
-        assert_eq!(
-            config.log_stdout_level, "info",
-            "Default should be used when LOG_LEVEL is unset"
-        );
-
-        // Then test with an empty string (same behavior as unset)
-        env_vars.insert("LOG_LEVEL", "");
-        let config = create_test_config_with_env(env_vars);
-        assert_eq!(
-            config.log_stdout_level, "info",
-            "Default should be used when LOG_LEVEL is empty"
-        );
-    }
-
-    #[test]
-    fn test_log_directory_mode_parsing() {
-        // Test the default value
-        let mut env_vars = HashMap::new();
-        env_vars.insert("ANTHROPIC_API_KEY", "test-api-key");
-
-        let config = create_test_config_with_env(env_vars.clone());
-        assert_eq!(
-            config.log_directory_mode,
-            LogDirectoryMode::Default,
-            "Default mode should be used when LOG_DIRECTORY_MODE is unset"
-        );
-
-        // Test explicit "default" value
-        env_vars.insert("LOG_DIRECTORY_MODE", "default");
-        let config = create_test_config_with_env(env_vars.clone());
-        assert_eq!(
-            config.log_directory_mode,
-            LogDirectoryMode::Default,
-            "Default mode should be used when LOG_DIRECTORY_MODE is 'default'"
-        );
-
-        // Test XDG mode
-        env_vars.insert("LOG_DIRECTORY_MODE", "xdg");
-        let config = create_test_config_with_env(env_vars.clone());
-        assert_eq!(
-            config.log_directory_mode,
-            LogDirectoryMode::Xdg,
-            "XDG mode should be used when LOG_DIRECTORY_MODE is 'xdg'"
-        );
-
-        // Test System mode
-        env_vars.insert("LOG_DIRECTORY_MODE", "system");
-        let config = create_test_config_with_env(env_vars.clone());
-        assert_eq!(
-            config.log_directory_mode,
-            LogDirectoryMode::System,
-            "System mode should be used when LOG_DIRECTORY_MODE is 'system'"
-        );
-
-        // Test case insensitivity
-        env_vars.insert("LOG_DIRECTORY_MODE", "XDG");
-        let config = create_test_config_with_env(env_vars.clone());
-        assert_eq!(
-            config.log_directory_mode,
-            LogDirectoryMode::Xdg,
-            "XDG mode should be used when LOG_DIRECTORY_MODE is 'XDG' (uppercase)"
-        );
-
-        // Test invalid value (should default)
-        env_vars.insert("LOG_DIRECTORY_MODE", "invalid");
-        let config = create_test_config_with_env(env_vars);
-        assert_eq!(
-            config.log_directory_mode,
-            LogDirectoryMode::Default,
-            "Default mode should be used when LOG_DIRECTORY_MODE is invalid"
-        );
-    }
-
-    #[test]
-    fn test_edge_case_unusual_path() {
-        // We'll use the create_test_config_with_env function directly, which properly
-        // manages environment variable state
-        let env_vars = HashMap::from([
-            ("ANTHROPIC_API_KEY", "test-api-key"),
-            ("LOG_FILE_PATH", "/dev/null/unusual/../path.log"),
-        ]);
-
-        // This function already handles setting and restoring environment variables safely
-        let config = create_test_config_with_env(env_vars);
-
-        // The issue happens on Linux where it doesn't properly use the LOG_FILE_PATH value
-        // Instead of expecting exact paths, we'll check that the path contains our unusual path components
-        // This works around platform-specific path handling while still testing the core functionality
-        assert!(
-            config.log_file_path.contains("unusual"),
-            "Path '{}' doesn't contain 'unusual'",
-            config.log_file_path
-        );
-    }
+    // Tests have been moved to tests/config_test.rs
 }
